Fix to release audio focus after call termination.
In cases where audio connection to bluetooth fails,
statemachine never transits to ActiveBluetoothRoute,
so when a call gets terminated, NO_FOCUS event is not
handled in QuiescentBluetoothRoute and audio focus is
not released.

Fix: Release audio focus by handling NO_FOCUS event,
even if the statemachine is in QuiescentBluetoothRoute.

Change-Id: I84ae1cc23e20e5c83fd091d233590d2515bfd1e0
CRs-Fixed: 2902314